Add HomoLR ML parts

+++ python/fate/ml/evaluation/classification.py
@@ -43,7 +43,7 @@ class MultiAccuracy(Metric):
 
     def __call__(self, predict, label, **kwargs) -> Dict:
         predict = self.to_np_format(predict, flatten=False)
-        label = self.to_np_format(label)
+        label = self.to_np_format(label).astype(np.int32)
         if predict.shape != label.shape:
             predict = predict.argmax(axis=-1)
         acc = accuracy_score(label, predict)
